计算机默认存储是小端，所以构造发送长度用int不会出错，但是会有隐患，尽量用short。

+++ day10-ProtobufServer/day10-AsyncServer/CSession.h
@@ -18,7 +18,7 @@ public:
 	MsgNode(const char * msg, short max_len):_total_len(max_len + HEAD_LENGTH),_cur_len(0){
 		_data = new char[_total_len+1]();
 		//×ªÎªÍøÂç×Ö½ÚÐò
-		int max_len_host = boost::asio::detail::socket_ops::host_to_network_short(max_len);
+		unsigned short max_len_host = boost::asio::detail::socket_ops::host_to_network_short(max_len);
 		memcpy(_data, &max_len_host, HEAD_LENGTH);
 		memcpy(_data+ HEAD_LENGTH, msg, max_len);
 		_data[_total_len] = '\0';
